server.registry.register
