+++ packages/omu/src/omu/extension/registry/registry.py
@@ -4,10 +4,42 @@ import abc
 from dataclasses import dataclass
 from typing import Callable
 
-from omu.extension.registry.packets import RegistryPermissions
-from omu.helper import Coro
+from omu.helper import Coro, map_optional
 from omu.identifier import Identifier
+from omu.network.bytebuffer import ByteReader, ByteWriter, Flags
 from omu.serializer import Serializable, Serializer
+
+
+@dataclass(frozen=True)
+class RegistryPermissions:
+    all: Identifier | None = None
+    read: Identifier | None = None
+    write: Identifier | None = None
+
+    def serialize(self, writer: ByteWriter) -> None:
+        flags = Flags(0, 3)
+        flags.set(0, self.all is not None)
+        flags.set(1, self.read is not None)
+        flags.set(2, self.write is not None)
+        writer.write_flags(flags)
+        if self.all is not None:
+            writer.write_string(self.all.key())
+        if self.read is not None:
+            writer.write_string(self.read.key())
+        if self.write is not None:
+            writer.write_string(self.write.key())
+
+    @classmethod
+    def deserialize(cls, reader: ByteReader) -> RegistryPermissions:
+        flags = reader.read_flags(3)
+        all_id = reader.read_string() if flags.has(0) else None
+        read_id = reader.read_string() if flags.has(1) else None
+        write_id = reader.read_string() if flags.has(2) else None
+        return RegistryPermissions(
+            map_optional(all_id, Identifier.from_key),
+            map_optional(read_id, Identifier.from_key),
+            map_optional(write_id, Identifier.from_key),
+        )
 
 
 @dataclass(frozen=True)
@@ -50,6 +82,10 @@ class RegistryType[T]:
 
 
 class Registry[T](abc.ABC):
+    @property
+    @abc.abstractmethod
+    def value(self) -> T: ...
+
     @abc.abstractmethod
     async def get(self) -> T: ...
