+++ packages/server/src/omuserver/extension/i18n/i18n_extension.py
@@ -1,3 +1,5 @@
+from omu.extension.i18n.i18n_extension import I18N_LOCALES_REGISTRY_TYPE
+
 from omuserver.server import Server
 
 from .permissions import (
@@ -12,3 +14,4 @@ class I18nExtension:
             I18N_GET_LOCALES_PERMISSION,
             I18N_SET_LOCALES_PERMISSION,
         )
+        server.registry.register(I18N_LOCALES_REGISTRY_TYPE)
