+++ packages/server/src/omuserver/extension/server/server_extension.py
@@ -52,7 +52,7 @@ class ServerExtension:
             REQUIRE_APPS_PACKET_TYPE,
         )
         server.permissions.register(SERVER_SHUTDOWN_PERMISSION)
-        self.version_registry = self._server.registry.create(VERSION_REGISTRY_TYPE)
+        self.version_registry = self._server.registry.register(VERSION_REGISTRY_TYPE)
         self.apps = self._server.tables.register(APP_TABLE_TYPE)
         server.network.listeners.connected += self.on_connected
         server.network.listeners.disconnected += self.on_disconnected
